Added a new element: flumpegshifterbin
The idea is to instantiate tsparse and flumpegshifter within the bin,
handle PMT message from tsparse and set pcr-pid property of the
timeshifter. This encapsulates the details about the interactions
between the timeshifter and tsparse elment.

The flumpegshifterbin currently supports only two element specific
properties which are simply forwarded to the timeshifter child
element:
- cach-size
- recording-template

Changed version to 0.10.1.2 and updated ChangeLog and README