Fixed issue with buffer references
The ringbuffer holds a list of buffers used as memory containers (slots).
When a buffer is sent dowstream, it's actually supposed to be consumed and
disposed. To avoid that and reuse buffers need to copy them before sending
them downstream. The copy is a "shallow" copy though so the actual memory
is shared.