Added a simple Python GUI app for playing videos
This app is inspired by the examples/pvr.py (which itself is based on
existing examples in Python bindings) but is much simpler and completely
rewritten to support GStreamer-1.0 and GObject introspection (manually
written bindings for GStreamer are no londger supported).

It's main purpose is to test the timeshifter element. Currently it plays by
default (no arguments) a pipeline which receives raw TS over UDP on port
10000.
If an HTTP URL is given on the command line as an argument, it will play
the pointed asset instead.