flutsbase: Remove virtual function update_segment
Mapping from BYTES to TIME can be accomplished solely using the index
which the base class has access to.  This means we don't need to call
the any derived class virtual functions and makes seeking, and thus
the src side of the element more independent from the sink (indexing)
side.