Add timeshiftseeker element
The timeshiftseeker element is part of an effort to split the
timeshifter into multiple elements such that the indexing/seeking/
ringbuffer can be reused independently.  It is currently an empty
shell which just passes buffers and events through verbatim but
code will be moved into it.

timeshiftseeker will be responsible for doing TIME->BYTES conversion
on upstream seek events and BYTES->TIME conversion on downstream
segment events based upon an index generated externally to the
element.