Remove index writer IDs from index API
GstIndex has a concept of "writer IDs".  My best guess is that back
in the day a single index would be shared between all streams.  This
is not the case in the timeshifter.  We share indices between
different elements by having our containing bin element
(mpegshifterbin) create the index and set it on each of our elements
as properties.  Writer IDs get in the way of this as to do real
sharing we would also have to pass a writer ID in as a property.

This commit removes the concept of writer IDs from the index class
by storing the ID in the index class itself.  This is not a real
solution as all of the implementation complexity is still present in
the index class.  Longer term I would like to replace the index
class with something much less complex so I don't fancy doing
anything about this right now.