tstimeshifter: Added support for buffering query
The tstimeshifter elment doesn't provide any property that would tell the
buffer size.  One of the reasons is the fact that the functionality is
split into separate elements: indexer, ring buffer and seeker.  This
makes it impossible for a signle element to convert byte offsets into
stream time and vice versa.

Buffering query adresses this problem as it allows elements in the pipline
to collect required information and cooperate.  It also seems to be
a more versitile approach.

* Refactored most of gst_ts_seeker_get_last_time() into
  gst_ts_seeker_bytes_to_stream_time()
  The latter one in an "accurate" mode might be grumpier (print warnings)

* added Pause button and buffering stats to the example video player