Signalling overrun in the timeshifter
The current behaviour is that whenever there's no space left in the ring
buffer, the "overrun" signal is emitted by the gsttstimeshifer element and
any incoming buffers are leaked.

It's up to the application to react to the "overrun" signal accordingly and
start popping buffers out of the ring buffer.  The first buffer popped
after an overrun is marked as discontinuous.