tsseeker:  Clamp byte offset lookups falling before the first indexed byte offset
This is an edge case when the first buffer that is indexed (the first one
with PCR) is not the first one in the stream.  This is quite common
situation as the PCR is included in only a fraction of all TS packets.

The upstream part of the pipeline may respond to a seek request for time 0
with a segment starting with a buffer before the first one the indexer has
ever seen.  The seeker should be tolerant to this situation as convert
segment start (an quite unlikely stop) time to 0.

The issue was reported due to segments pushed with invalid start time
for seek-to-time-0 requests.