tsshifter:  Fix memory leak when finalizing the timeshifter
The tsshifter element owns the ring buffer (tscache) and should unref it
when finalizing.  The memory that leaked was a set of buffers allocated for
the ring buffer.

The ring buffer is disposed correctly after flushing.  The missing
deallocation when finalizing the timeshifter haven't been noticed as
usually the process that owned the pipeline ceases to exist.  In the real
scenario though the application process still exists but simply creates
a new pipeline (e.g. after tuning to a new channel) but the memory is
already leaked.