To increase performance remove tsparse element.
To increase performance of the timeshifter remove tsparse element
from the bin after it's no longer needed.

tsparse is only needed to obtain PID of the stream we're interested in.
Once this is done we don't need to parse the stream. By removing
tsparse it's possible to lower the CPU usage by a few percent
(relative to the other elements in the tsshifterbin).

To make removal of element in a live pipeline possible, a fix had to
be applied in tsshifter, because it was dropping caps event which
was upsetting upstream elements (non-negotiated error).