Timeshifter: Avoid trashing caches and degrading performance
The timeshifter writes incoming mpeg-ts data to disk and reads it back
out again.  This involves a lot of disk I/O which goes via the linux
kernel page-cache.

The kernel caches pages in memory as it usually improves performance.
Unfortunately, due to limited resources, the kernel must drop some pages
from the cache to let others in.  If the kernel gets it wrong it may
evict some pages which are required in favour of pages which are not
likely to be used again.  We have seen this cause performance degredation
when an channel is tuned to for a long time.  This is because some of the
code pages required for other uses have been evicted from cache and must
be read again from disk.

This change modifies the timeshifter to give the kernel more information
about the expected use of the pages it uses.  There are two hints given:

1. When a timeshifter slot is filled with mpeg-ts data the timeshifter
   will now tell the kernel to start writing it to disk such that the
   page will become "clean" rather than "dirty".
2. After giving the kernel some time to asynchronously write it to disk
   the writer will ensure that it has been written (blocking if
   necessary) and instruct the kernel to drop the page from cache.
3. We also drop the page if it has been read as it it unlikely to be read
   again.

You can measure which pages are resident in memory with linux-ftools[1].
With this it's easy to see the difference in the number of pages of the
timeshift-buffer are resident in memory.

With this patch there is no performance degredation when as the
timeshifter is used.

[1]: http://code.google.com/p/linux-ftools/