Fix a problem when a PCR would wrap causing the timeshifter to fail
PCR or Program Clock Reference is a 90Khz clock that is embedded into
mpegts packets. Timeshifter uses this clock to bind stored buffers
and they time of arrival for future reference. One of the properties
of PCR clock is that it is stored as a 33 bit value hence it will wrap
every ~26.5 hours. Timeshifter would previously fail in such a case
because stream time was directly derived from the arriving PCR and if
discontinuity would be encountered timeshifter would fail to properly
account arriving buffers. Also buffer level would be incorrectly flagged.

This change amends that by not using PCRs as direct time reference.