Implement an element that will aid in PCR wraping tests.
Timeshifter has to be robust against PCR wraps. This element
can manipulate PCRs of the mpegts stream such that it is easy to produce
a situation where such a wrap occurs.

This is only meant to be used in tests for timeshifter element.
This is because one of the limitations of tsstamper is the fact
that only restamps PCRs and *NOT* PTSs, DTSs, SCRs etc.
This will effect in a stream that is confusing for the decoder.
For this reason it is only useful in tests of timeshifter where
we don't do any decoding.