Fix memory leak due to improper refcounting
This commit fixes a refcounting error, causing a significant memory
leak.

The error was caused by gsttsseeker and gsttsindexer set_property
function calling `g_value_dup_object` which increases the refcount. This
was sufficient to preserve the ownership, but the erroneous code was
increasing the refcount once more. This has been corrected and it
appears that memindex is properly cleaned up.