Reduce memory usage when using large backing files
For easy tracking of what data has been written where the timeshifter
divides its backing file store into slots, each 32kB in size.  It keeps
accounting information in memory for each one of these slots in RAM as
a `struct Slot`.  This in turn contains a pointer to a GstBuffer which
corresponds to that slot.

Unfortunately this means that timeshifter memory usage scales linearly
with backing store size.  For realistic sizes of the backing store for
storing 2 hours of HD video (~10GB) this memory overhead can be on the
order of ~100MB.  This is too much.

This commit fixes this issue by reducing this per slot overhead by not
storing a `GstBuffer` for each slot.  Instead `GstBuffer`s are created
on demand on the read side and not used at all on the write side.  This
has necessatated a change in abstraction.  Previously the timeshifter
wasn't even directly aware that it was backed by a file[1].  A Gstreamer
`GstAllocator` was injected into it which in this case was a
`GstFileMemAllocator`, requiring a patch adding this to the gstreamer
sources.

Instead now a file descriptor is injected into the timeshifter for use
as a timeshift buffer.  This provides a different kind of flexibility
to the model of injecting a `GstAllocator`.  This patch also means that
there is no longer a dependency on `GstFileMemAllocator` at all so the
timeshifter can be built against pristine Gstreamer.

With this patch memory usage still scales linearly with backing store
size, just less severely.  Future work could involve removing `struct
Slot` entirely to reduce memory usage further.

[1]: This is not strictly true.  The introduction of the efforts to avoid
     cache trashing actually made timeshifter directly dependant on the
     fact that the `GstFileMemAllocator` was used to allocate these
     buffers.